v3.0.0 The first steps of class 'AzBlobServiceClientExtensions' has been implemented, containing scaffolding methods to perform operations with azure storage blobs.
Installed dependencies:
Azure.Storage.Blobs.12.10.0
Azure.Storage.Common.12.9.0

To better understand and avoid confusion, all the methods in the class 'AzTableClientExtensions' with the response 'AzStorageResponse <AsyncPageable<T>>' have been renamed starting with the text 'AsynPageable' and continuing with the main text. The above methods in the class 'AzTableClientExtensions' that ends with the text 'AsyncPageable' have been declared obsolete.